requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.0
django-cors-headers==4.3.1
django-debug-toolbar==5.2.0
django-filter==23.3
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
python-slugify==8.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==80.7.1
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.6.0
